Render chart annotations correctly when data is grouped
Chart annotations only worked when the chart plots a single data
series. If the chart used the “Group By” option, the calculation of
which point on the X-axis to render the annotation at was incorrect,
causing the selector for the chart element to return `undefined`.

There doesn’t seem to be away to access the chart options, so I had to
add a `group_by` option to the `chartAnnotations()` function, which
should mirror the Group By setting of the chart.