Handle url specials from cell data in addLinksToTables
When the cell contents includes URL-special characters, they should probably
not be interpreted as multiple query parameters or other such errors.

For example, with a addLinksToTables({link_columns: ["col"], link_urls: ["http://modeanalytics?param_foo={{col}}"]}):
  Cell Contents: `A+B&securityhole=true`
  Before: `http://modeanalytics/?param_foo=A+B&securityhole=true`
  After: `http://modeanalytics/?param_foo=A%2bB%26securityhole=true`

So this avoids data that is potentially sourced from external users getting
injected into the super dangerous "securityhole" parameter on another report
unintentionally.

Or more practically, when the data contains a +, it doesn't get re-interpreted
as a space by the browser.

IMO, interpreting data from a single cell in a table as multiple parameters
should be a special case that needs more coding. The simple case of expecting
the cell contents to be a single scalar value should be the default.